requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
moviepy>=1.0.3
scipy>=1.7.0
pyyaml>=5.4.0
pytest>=6.2.0
python-logging-config>=0.5.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
openai-whisper>=20231117
jieba>=0.42.1
scikit-learn>=1.0.0
ffmpeg-python>=0.2.0
nltk>=3.8.1
